requirements.txt
flask
flask-cors
mahotas
numpy
tifffile
matplotlib
pillow